+++ R/run_baseline_nowcast_pipeline.R
@@ -31,10 +31,6 @@
 #'    `long_df_for_borrow` to estimate the dispersion parameters, default is
 #'    `FALSE`.
 #' @autoglobal
-#' @importFrom baselinenowcast add_obs_errors_to_nowcast
-#'    aggregate_df_by_ref_time generate_point_nowcasts
-#'    generate_triangles get_delay_estimate estimate_dispersion
-#'    truncate_triangles apply_delay nowcast_list_to_df
 #' @importFrom dplyr select distinct filter mutate arrange left_join row_number
 #'    bind_rows pull
 #' @importFrom tibble tibble
@@ -137,10 +133,10 @@ run_baselinenowcast_pipeline <- function(long_df,
 #'    `FALSE`.
 #'
 #' @autoglobal
-#' @importFrom baselinenowcast add_obs_errors_to_nowcast
-#'    aggregate_df_by_ref_time generate_point_nowcasts
+#' @importFrom baselinenowcast add_uncertainty
+#'    aggregate_df_by_ref_time generate_pt_nowcast_mat_list
 #'    generate_triangles get_delay_estimate estimate_dispersion
-#'    truncate_triangles apply_delay nowcast_list_to_df
+#'    truncate_triangles apply_delay nowcast_matrix_list_to_df
 #' @importFrom dplyr select distinct filter mutate arrange left_join row_number
 #'    bind_rows pull
 #' @importFrom tibble tibble
@@ -188,14 +184,14 @@ baselinenowcast_pipeline <- function(long_df,
   }
 
   delay <- get_delay_estimate(
-    triangle = triangle_for_delay,
+    reporting_triangle = triangle_for_delay,
     max_delay = max_delay,
     n = n_history_delay
   )
 
   # Get point estimate
-  point_reporting_square <- apply_delay(
-    triangle_to_nowcast = triangle,
+  point_nowcast_mat <- apply_delay(
+    rep_tri_to_nowcast = triangle,
     delay = delay
   )
 
@@ -211,26 +207,28 @@ baselinenowcast_pipeline <- function(long_df,
   } else {
     triangle_for_uncertainty <- triangle
   }
-
-  truncated_rts <- truncate_triangles(triangle_for_uncertainty,
+  truncated_rts <- truncate_triangles(
+    reporting_triangle = triangle_for_uncertainty,
     n = n_history_uncertainty
   )
   # Get retrospective triangles
-  retro_rts <- generate_triangles(list_of_trunc_rts = truncated_rts)
+  retro_rts <- generate_triangles(trunc_rep_mat_list = truncated_rts)
   # Get retro nowcasts
-  retro_nowcasts <- generate_point_nowcasts(list_of_rts = retro_rts)
+  retro_nowcasts <- generate_pt_nowcast_mat_list(
+    reporting_triangle_list = retro_rts
+  )
   disp_params <- estimate_dispersion(
-    list_of_nowcasts = retro_nowcasts,
-    list_of_trunc_rts = truncated_rts
+    pt_nowcast_mat_list = retro_nowcasts,
+    trunc_rep_mat_list = truncated_rts
   )
 
-  exp_obs_nowcasts <- add_obs_errors_to_nowcast(
-    comp_rep_square = point_reporting_square,
+  exp_obs_nowcasts <- add_uncertainty(
+    point_nowcast_matrix = point_nowcast_mat,
     disp = disp_params,
     n_draws = n_draws
   )
-  nowcast_draws <- nowcast_list_to_df(
-    exp_obs_nowcasts
+  nowcast_draws <- nowcast_matrix_list_to_df(
+    nowcast_matrix_list = exp_obs_nowcasts
   )
   ind_nowcast <- aggregate_df_by_ref_time(nowcast_draws)
   # Join data and observed data to this!